💰 This starts the second version "bible-overtype"
This is a new codebase to start over again to code
a new version of the same program but this time it
will be called just "bible" (not "bible-overtype",
because in future we will be adding more features)
and it will be a fresh rewrite avoiding the issues
which previous version has. The old version is not
very well designed, it does not follow any of good
practices of C coding, its JS code is just brought
out of other projects and is not maintainable. Now
we are starting this new version that will have no
such issues, the new version will have no parts of
code that is not specifically written for this new
project, also it will be WASM first, and Linux CLI
app second. It will only have minimal JS code that
will implement curses functionality for WASM build
usin Xterm.js. And the C code will be followin all
the good practices like freeing memory, it will be
using safe memory access, and will be well tested.