🧱 Build script for Linux command line application
Provides Docker container setup to build the Linux
version of the Bible application and run it on any
operating system in Alpine container. On top of it
main function in also a part of this commit, as it
is only used in CLI application. Note that the web
version of Bible does not call this main function.