🚧 Both Alpine and WASM builds can use same colors
Ported curses routines attron, attroff, addstr and
others, related to setting the color pairs and use
them while writing text to the screen. Not all the
functions are implemented, some just took a simple
way to make the demo application work. But soom we
will enhance them. Just wait for it... Also in the
same commit is the main implementation which makes
program life-cycle menthod to be called in managed
way so that both Alpine using this make method and
Web application using WASM behaves in exactly same
way, because main methlod logic mirrors the JS app
closely and the behavior is more or less the same.