✨ Add PDCurses from Emcurses project for WASM app
Wow, I finally got the library show clear signs of
working with Xterm.js. I need to remind that stock
Emcurses uses termlib.js - which is a very old and
not maintained library. Javascript world has a new
champion for terminal, and it is Xterm. So, I have
forked Emcurses project and changed it so that not
only it does not depend on Termlib, but actually I
left for library integrators to write a Javascript
glue code so that it could be easily switched from
one frontend, like our Xterm.js, to any technology
which we might want to use for rendering terminal.