➕ Use PDCurses latest version instead of Emcurses
Emcurses is an awesome piece of work without it, I
could have never built the first iteration of this
application (which is called bible-overtype, which
some of you know already). And I also started this
version of Bible application based realy on a fork
of Emcurses (Just changed it to use Xterm.js, many
of you I am sure still remember this). Sadly, this
Emcurses project has not been developed since like
ten years by now, and my changes also require some
sort of fork, and since Emcurses is a for of other
library already (PDCurses), I would rather fork of
the original version, rather the fork of it. Since
then PDCurses is itself is on GitHub, and it is at
version 3.9 already (it is what has been released,
and the code in git is even later that that) while
Emcurses is based on v3.4 of PDCurses. I just took
the code in emcurses/emscripten and placed it into
./wasm in PDCurses directory, then I did some tiny
adjustments to make it build, and now I have sweet
new fork of PDCurses. Emcurses howevere never felt
like it it was well tested, and "window" functions
were not quite working, sadly moving into PDCurses
did not make it work, so i will be looking in into
how can I improve my fork to support Curses fully.