🐛 Fix the bug which occured while drawing windows
Window functions did not work for WASM build until
this commit, fixing the bug. In curses JS callback
it was ignoring row and col parameters at all this
way not moving cursor while drawing a window, thus
all the window output happend to be written in one
line. The fix it bug involves moving cursor to the
position where text output was expected to happen.