🌐 Display emojis and UTF-8 characters on PDCurses
The work was mostly carried out in PDCurses, where
it was introduced another callback which allows to
print entire string without putting color codes in
between the bytes of a single UTF-8 character. Yet
we don't have the character width support it still
counts bytes to determine the length of the sring,
what is considered bug, and will have to be fixed!