🚧 Print emojis after refresh as a workaround flow
Emojis (and other multi-byte characters) were here
causing some screen artifacts, due to their length
in bytes curses would allocate more spaces to show
them, compared to what was actually needed, making
other characters shift and leave space between the
emoji and the anticipated start of the rest of the
line text. This commit is not really solving these
issues, but provides an example workaround however
this depends on refresh, therefore: not efficient!