✨ Full support for 16 bit Unicode Emoji character
What could possibly be encoded as a 16 bit Unicode
is now fully supported. However, support for wider
Unicodes that are above 0xFFFF is not here yet (as
this exceeds current PDCurses 16bit implementation
of wide character sadly. But a lot of Emojis right
now are supported already, as they fit in 2 bytes!