🧵 We embrace the async nature of WASM event loop!
Introduce while_getch routine which combines loops
through while construct and getch call inside. The
reason for that is to provide a way to refactor it
using this primitive which works well both on that
old curses for terminal and this new version which
provides web experience. The only difference right
now is the fact that NCurses version is not really
async, it is actually blocking. But it is not that
big deal as there should never be tho much of code
after a call that starts the event loop. Let's now
just hope that I could find some time to fix this.