🐬 Fix Alpine build / Separated channel for resize
Resize is signaled by KEY_RESIZE in NCurses but in
WASM we don't use getch, and Xterm.js natively can
send a message to the WASM program, the payload of
which includes the dimensions of the current size.

Also added some re-setting terminal properties and
global variable COLS and ROWS to match the message
of the event which occurs when the window resizes.