🧵 Asyncify with standalone library makes blocking
Finally, managed to fetch some luck with Asyncify,
a standalone library without Emscripten spaghetti,
which provides just what I need, and I want to say
that there is a working example of a blocking call
but everything needs cleanup, to be in right place
and to make sure it does not break Alpine version.

This is just an experiment, but it is realy a game
changer! If you look at the previous commits, then
it will become apparent that this was an issue for
a while. Finally I was ready to settle with just a
way to use curses in async way only. Doing it this
way would enable me to continue workin of the main
Bible program, but it would make this solution not
re-usable, outside of the Bilble project. And here
in this commit there is a proof of concept for the
Asyncifying, which will allow to work on fully and
completelly functional port of PDCurses into WASM!