🔖 Events loop is rebuilt with synchronous timeout
Finally async-only days are over! Here we can even
look into porting PDCurses to WASM fully. However,
hanging to this reveals that I have been using the
functions from the curses library wrong. Here with
the loop refreshing stdscr, there is some flicker,
but it is mostly because I should have have better
not using getch for entire stdscr, wgetch function
would be better, so that it only refreshes the one
window in which we are looking. This problem never
obvious before but it always existed on the Alpine
version. It became apparent because slow rendering
in the browser. But it is good to ralize we really
have that issue so that somebody could fix it now.