💩 Handle screen resizes in bit hacky way for WASM
After introducing blocking for getch() effectively
it is blocked at nearly all times, and in is never
a good time to call any of module.instance.exports
provided functions, to prevent reentrancy. On it's
bright side, it gives an opportunity to have there
the same resize code for both Alpine and WASM. But
now some quick patches to WASM PDcurses are added.