🩹 Dirty setup to get libunistring on WASM working
The project builds, but started to crash on bigger
files. Also `apt installlibc6-dev-i386` is failing
in the Dockerfile, hopefully it is ignored, but on
manual attempt to install it seems the project now
compiles, but this all need to cleanup, just it is
time consuming and I just wanted to save progress.