🚨 compiler warnings erradicated, also build fixed
There were a lot of warnings reported for WASM 32b
architecture, because on it char* and uint8_t does
not match and therefore explicit cast statement is
needed. This commit also adds a fix for the header
which relates to 32 b architecture but that header
is always required by another libunistring header.