🎉 Do not use my custom fork from PDCurses anymore
First commit of 2023, adds all changes in PDCurses
and including platform for WASM into this repo, so
that it uses stock PDCurses. Later, of course, the
files from pdcurses in this repo will be attempted
to merge back into upstream if only they accept it
at all. Anyway, happy new year, and be happy kids!